Work on processors in progress

+++ src/satmo/preprocessors.py
@@ -5,13 +5,15 @@ import os
 import bz2
 import glob
 import warnings
+import string
 
-from .utils import super_glob, parse_file_name
+from .utils import super_glob, parse_file_name, make_file_path, is_day
 from .errors import SeadasError
+from .global_variables import PRODUCT_SUITES
 
 
 
-def bz2_unpack(source, destination):
+def bz2_unpack(source, destination, overwrite = False):
     """Unpacks data compressed with bz2
     
     Utility function to unpack bz2 compressed data
@@ -21,6 +23,7 @@ def bz2_unpack(source, destination):
         source (str): the filename of the archive
         destination (str): the destination directory, filename will be
         set automatically
+        overwrite (bool): If false, check whether file already exists and returns its name
 
     Returns:
         str: The filename of the unpacked file
@@ -28,13 +31,15 @@ def bz2_unpack(source, destination):
     if not os.path.exists(destination):
         os.makedirs(destination)
     out_file = os.path.join(destination, os.path.splitext(os.path.basename(source))[0])
+    if not overwrite and os.path.isfile(out_file):
+        return out_file
     with open(out_file, 'wb') as dst, bz2.BZ2File(source, 'rb') as src:
         for data in iter(lambda : src.read(100 * 1024), b''):
             dst.write(data)
     return out_file
 
 
-def bz2_compress(source, destination, compresslevel = 3):
+def bz2_compress(source, destination, compresslevel = 3, overwrite = False):
     """Unpacks data compressed with bz2
     
     Utility function to bz2 compress data
@@ -44,6 +49,7 @@ def bz2_compress(source, destination, compresslevel = 3):
         destination (str): the destination directory, filename will be
         set automatically
         compresslevel (int): Compression level
+        overwrite (bool): If false, check whether file already exists and returns its name
 
     Returns:
         str: The filename of the bz2 compressed file
@@ -51,6 +57,8 @@ def bz2_compress(source, destination, compresslevel = 3):
     if not os.path.exists(destination):
         os.makedirs(destination)
     out_file = os.path.join(destination, os.path.basename(source)) + '.bz2'
+    if not overwrite and os.path.isfile(out_file):
+        return out_file
     with open(source, 'rb') as src, bz2.BZ2File(out_file, 'wb', compresslevel = compresslevel) as dst:
         for data in iter(lambda : src.read(100 * 1024), b''):
             dst.write(data)
@@ -60,15 +68,52 @@ def bz2_compress(source, destination, compresslevel = 3):
 # Make a class that holds all the information of a future L2 (binned) product
 
 class l3mapProcessJob(object):
-    def __init__(self, input_dir, pattern = r'.*'):
+
+    template = Environment(loader=PackageLoader('satmo', 'templates')).get_template('l3bprocess.par')
+
+    def __init__(self, input_dir, pattern = r'.*', output_basedir = None):
         self.input_dir = input_dir
         self.file_list = super_glob(input_dir, pattern)
-    def execute(self, north, south, east, west):
-        status = subprocess.call(['multilevel_processor.py', 'modis_SR.par'])
-        # Status should be zero if command completed successfully
+        self.sensor = parse_file_name(self.file_list[0])['sensor']
+        if output_basedir is None:
+            output_basedir = string.replace(input_dir, make_file_path(self.file_list[0], add_file = False), '')
+        self.output_dir = os.path.join(output_basedir, make_file_path(self.file_list[0], doy = False, level = 'L3b'))
+
+    def execute(self, north, south, east, west, suite, day = True, binning_resolution = 1,\
+        mapping_resolution = '1km', proj4 = '+proj=laea +lat_0=18 +lon_0=-97', overwrite = False):
+        # Day or night only filter
+        file_list_sub = [x for x in self.file_list if is_day(x) is day]
+        # Extract files if they need extraction
+        ext = os.path.splitext(file_list_sub[0])[1]
+        if ext == '.bz2':
+            file_list_sub = [bz2_unpack(x, self.input_dir) for x in file_list_sub]
+        # Create output dir if it doesn't exist
+        if not os.path.exists(self.output_dir):
+            os.makedirs(self.output_dir)
+        # Create variables for templating and command run 
+        overwrite_int = int(overwrite)
+        prod_list = PRODUCT_SUITES[suite][self.sensor]
+        # Par file templating 
+        par = self.template.render(binning_resolution = binning_resolution,
+                                   overwrite_int = overwrite_int,
+                                   prod_list = prod_list,
+                                   file_list = file_list_sub)
+        # par filename creation + writing to output directory
+        par_file = os.path.join(self.output_dir, '%s_l3bprocess.par' % self.sensor)
+        with open(par_file, "wb") as dst:
+            dst.write(par)
+        # Build multilevel processing arguments list
+        arg_list_0 = ['multilevel_processor.py', par_file, '--use_existing', '--output_dir=%s' % self.output_dir]
+        status0 = subprocess.call(arg_list_0) # --use_existing --output_dir=  (--overwrite - also in par file)
+        if status0 != 0:
+            raise SeadasError('Multilevel processor exited with %d during processing to L2bin' % status)
+        # Prepare l3mapgen command and run it
+        # l3mapgen ifile=T2016292.L3b_DAY_OC ofile=T2016292.L3B_DAY_RRS_laea.nc 
+        # resolution=1km south=26 north=40 west=-155 east=-140 projection="+proj=laea +lat_0=33 +lon_0=-147"
+        return output_file
+
     def clean(self):
         pass
-
 
 
 class extractJob(object):
